RF: Shift former 'install' tests that are now 'add' tests
