+++ datalad/distribution/tests/test_install.py
@@ -208,21 +208,6 @@ def test_install_plain_git(src, path):
     ok_('test2.txt' in ds.repo.get_indexed_files())
 
 
-@with_tree(tree={'test.txt': 'some',
-                 'dir': {'testindir': 'someother',
-                         'testindir2': 'none'}})
-def test_install_files(path):
-    ds = create(path)
-    # install a single file
-    eq_(ds.install('test.txt'), opj(path, 'test.txt'))
-    # install it again, should given same result
-    eq_(ds.install('test.txt'), opj(path, 'test.txt'))
-    # install multiple files in a dir
-    eq_(ds.install('dir', recursive=True),
-        [opj(path, 'dir', 'testindir'),
-         opj(path, 'dir', 'testindir2')])
-    # TODO: check git
-
 
 @with_testrepos(flavors=['local-url', 'network', 'local'])
 @with_tempfile
@@ -256,55 +241,6 @@ def test_install_dataset_from_just_source_via_path(url, path):
     ok_(ds.is_installed())
     ok_clean_git(ds.path, annex=False)
     assert_true(os.path.lexists(opj(ds.path, 'test-annex.dat')))
-
-@with_testrepos(flavors=['local-url', 'network', 'local'])
-@with_tempfile
-def test_install_into_dataset(source, top_path):
-    ds = create(top_path)
-    subds = ds.install(path="sub", source=source)
-    assert_true(isdir(opj(subds.path, '.git')))
-    ok_(subds.is_installed())
-    # sub is clean:
-    ok_clean_git(subds.path, annex=False)
-    # top is not:
-    assert_raises(AssertionError, ok_clean_git, ds.path, annex=False)
-    # unless committed the subds should not show up in the parent
-    # this is the same behavior that 'git submodule status' implements
-    assert_not_in('sub', ds.get_subdatasets())
-    ds.save('addsub')
-    assert_in('sub', ds.get_subdatasets())
-
-
-@with_testrepos('submodule_annex', flavors=['local', 'local-url', 'network'])
-@with_tempfile(mkdir=True)
-def test_install_subdataset(src, path):
-    # get the superdataset:
-    ds = install(path=path, source=src)
-
-    # subdataset not installed:
-    subds = Dataset(opj(path, 'sub1'))
-    assert_false(subds.is_installed())
-
-    # install it:
-    ds.install('sub1')
-    assert_true(isdir(opj(subds.path, '.git')))
-
-    ok_(subds.is_installed())
-    # Verify that it is the correct submodule installed and not
-    # new repository initiated
-    assert_equal(set(subds.repo.get_indexed_files()),
-                 {'test.dat', 'INFO.txt', 'test-annex.dat'})
-
-    # Now the obnoxious install an annex file within not yet
-    # initialized repository!
-    with swallow_outputs():  # progress bar
-        ds.install(opj('sub2', 'test-annex.dat'))
-    subds2 = Dataset(opj(path, 'sub2'))
-    assert(subds2.is_installed())
-    assert(subds2.repo.file_has_content('test-annex.dat'))
-    # we shouldn't be able silently ignore attempt to provide source while
-    # "installing" file under git
-    assert_raises(FileInGitError, ds.install, opj('sub2', 'INFO.txt'), source="http://bogusbogus")
 
 
 @with_tree(tree={
